Add files via upload
Grade assessment of a student
The function of this project is to assess the grade of a student in an exam. If the student scores more than or
equal 90, he will achieve "A" . If the student scores more than 80 and less than 90 he will get "B". If the student
scores more than 70 and less than 80 he will get "C" . Finally, if he scores less than 50, he will get "F"